+++ src/run_file_ingest.py
@@ -11,7 +11,7 @@ from minio.commonconfig import Tags
 
 from timeio.common import get_envvar, setup_logging
 from timeio.databases import DBapi
-from timeio.errors import UserInputError, ParsingError
+from timeio.errors import UserInputError, ParsingError, ParsingWarning
 from timeio.feta import Thing
 from timeio.journaling import Journal
 from timeio.mqtt import AbstractHandler, MQTTMessage
@@ -74,21 +74,23 @@ class ParserJobHandler(AbstractHandler):
         rawdata = self.read_file(bucket_name, filename)
 
         logger.info("parsing rawdata ... ")
-        file = source_uri
-        with warnings.catch_warnings() as w:
+        file = "/".join(source_uri.split("/")[1:])  # remove bucket name from source_uri
+        with warnings.catch_warnings(record=True) as recorded_warnings:
+            warnings.simplefilter("always", ParsingWarning)
             try:
                 df = parser.do_parse(rawdata)
                 obs = parser.to_observations(df, source_uri)
             except ParsingError as e:
                 journal.error(
-                    f"Parsing failed. Detail: {e}. File: {file!r}", thing_uuid
+                    f"Parsing failed. File: {file!r} | Detail: {e}", thing_uuid
                 )
                 raise e
             except Exception as e:
-                journal.error(f"Parsing failed for file {file!r}", thing_uuid)
+                journal.error(f"Parsing failed. File: {file!r}", thing_uuid)
                 raise UserInputError("Parsing failed") from e
-            if w:
-                journal.warning(w[0].message, thing_uuid)
+            for w in recorded_warnings:
+                logger.info(f"{w.message!r}")
+                journal.warning(f"{w.message}", thing_uuid)
 
         logger.debug("storing observations to database ...")
         try:
@@ -102,8 +104,14 @@ class ParserJobHandler(AbstractHandler):
             )
             raise e
 
-        # Now everything is fine and we tell the user
-        journal.info(f"Parsed file {file}", thing_uuid)
+        if len(obs) > 0:
+            # Now everything is fine and we tell the user
+            journal.info(
+                f"Parsed file: {file!r} | "
+                f"Data rows: {df.shape[0]} | "
+                f"Stored observations: {len(obs)}",
+                thing_uuid,
+            )
 
         object_tags = Tags.new_object_tags()
         object_tags["parsed_at"] = datetime.now().isoformat()